add umbrella_defineopt and umbrella_opt_default to help manage umbrella config options.
umbrella option variables are built on top of cmake CACHE
and normal variables.  we put the basic definition of the
option variable in the umbrella/*.cmake file, but we allow
the user to override the variable's default value by loading
it into a normal variable and using that to define the
cache variable.   users can also directly set a normal
variable to override all the defaults (including -D from
the cmake command line).

umbrella_defineopt: define an umbrella option variable in
the cmake cache.

umbrella_opt_default: set a user-provided default value for an
umbrella option variable.